requirements.txt
# Core AI/ML Frameworks
langchain==0.1.14
langchain-core==0.1.33
langchain-community==0.0.34
langchain-google-genai==0.0.14
langchain-openai==0.1.6
llamaindex==0.9.46
langgraph==0.0.10
crewai==0.27.0

# Vector DB & Storage
qdrant-client==2.7.0
pinecone-client==3.0.2
python-dotenv==1.0.0

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Frontend
gradio==4.11.0

# Document Processing
pdf2image==1.16.3
pypdf==3.17.1
python-docx==0.8.11
openpyxl==3.10.10
python-pptx==0.6.21
BeautifulSoup4==4.12.2

# Web Scraping & Search
requests==2.31.0
tavily-python==0.1.0
selenium==4.15.2

# Utils
pyyaml==6.0.1
tenacity==8.2.3
typing-extensions==4.8.0
httpx==0.25.2
aiohttp==3.9.1

# Development
pytest==7.4.3
black==23.12.1
flake8==6.1.0